Add a dependency: flame
