+++ pubspec.yaml
@@ -35,7 +35,7 @@ dependencies:
   path_provider: ^2.1.4
   geolocator: ^13.0.1
   provider: ^6.1.2
-
+  flame: ^1.19.0
 
 
 
@@ -66,9 +66,8 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
+  assets:
+    - assets/images/
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/to/resolution-aware-images
